snapshot syncing refactor
