+++ .changeset/loud-lamps-jump.md
@@ -1,0 +1,5 @@
+---
+'backend': patch
+---
+
+moving snapshot syncing to a separate functions
