+++ modules/controllers/index.ts
@@ -3,3 +3,4 @@ export * from './snapshots-controller';
 export * from './pools-mutation-controller';
 export * from './user-balances-controller';
 export * from './cow-amm-controller';
+export * from './pool-events-query-controller';
