+++ worker/job-handlers.ts
@@ -15,10 +15,9 @@ import moment from 'moment';
 import { cronsDurationMetricPublisher } from '../modules/metrics/cron-duration-metrics.client';
 import { syncLatestFXPrices } from '../modules/token/latest-fx-price';
 import { AllNetworkConfigs, AllNetworkConfigsKeyedOnChain } from '../modules/network/network-config';
-import { JobsController } from '../modules/controllers/jobs-controller';
 import { chainIdToChain } from '../modules/network/chain-id-to-chain';
 import { Chain } from '@prisma/client';
-import { CowAmmController } from '../modules/controllers';
+import { JobsController, CowAmmController, SnapshotsController } from '../modules/controllers';
 
 const runningJobs: Set<string> = new Set();
 
@@ -238,7 +237,7 @@ export function configureWorkerRoutes(app: Express) {
                 await runIfNotAlreadyRunning(
                     job.name,
                     chainId,
-                    () => poolService.syncLatestSnapshotsForAllPools(),
+                    () => SnapshotsController().syncSnapshotsV2(chainId),
                     res,
                     next,
                 );
